+++ Editor/Playa/Renderer/PlayaSeparatorSemiRenderer/PlayaSeparatorRendererIMGUI.cs
@@ -6,17 +6,17 @@ namespace SaintsField.Editor.Playa.Renderer.PlayaSeparatorSemiRenderer
     {
         protected override void RenderTargetIMGUI(float width, PreCheckResult preCheckResult)
         {
-            throw new System.NotImplementedException();
+            throw new System.NotImplementedException("PlayaSeparatorRenderer is not supported in IMGUI. Try UI Toolkit instead");
         }
 
         protected override float GetFieldHeightIMGUI(float width, PreCheckResult preCheckResult)
         {
-            throw new System.NotImplementedException();
+            throw new System.NotImplementedException("PlayaSeparatorRenderer is not supported in IMGUI. Try UI Toolkit instead");
         }
 
         protected override void RenderPositionTargetIMGUI(Rect position, PreCheckResult preCheckResult)
         {
-            throw new System.NotImplementedException();
+            throw new System.NotImplementedException("PlayaSeparatorRenderer is not supported in IMGUI. Try UI Toolkit instead");
         }
     }
 }