+++ Runtime/Interfaces/ISeparatorAttribute.cs
@@ -1,9 +1,11 @@
+using UnityEngine;
+
 namespace SaintsField.Interfaces
 {
     public interface ISeparatorAttribute
     {
         string Title { get; }
-        EColor Color { get; }
+        Color Color { get; }
         EAlign EAlign { get; }
         bool IsCallback { get; }
         int Space { get; }